+++ src/containers/Blog/Blog.js
@@ -1,50 +1,11 @@
 import React, { Component } from 'react';
 //import axios from 'axios'
-import axios from '../../axios'
 import './Blog.css';
+import Posts from './Posts/Posts'
+import { Route } from 'react-router-dom'
 
 class Blog extends Component {
-    state = {
-        posts: [],
-        selectedPostId: null,
-        error: false
-    }
-
-    componentDidMount () {
-        axios.get('https://jsonplaceholder.typicode.com/posts')
-            .then(response =>{
-                const posts = response.data.slice(0, 4);
-                const updatedPosts = posts.map(post => {
-                    return {
-                        ...post,
-                        author: 'Max'
-                    }
-                });
-                this.setState({posts: updatedPosts})   
-                //console.log(response);
-            })
-            .catch(error => {
-                //console.log(error);
-                this.setState({error: true})   
-            }) 
-    }
-
-    postSelectedHandler = (id) => {
-        this.setState({selectedPostId: id});
-    }
-
-    render () {
-        let posts = <p style={{textAlign: "center"}}>something went wrong!</p>
-        if (!this.state.error) {
-            posts = this.state.posts.map(post => {
-                return <Post 
-                    key={post.id} 
-                    title={post.title} 
-                    author={post.author}
-                    clicked={()=>this.postSelectedHandler(post.id)} />
-            });
-        }
-        
+    render () {  
         return (
             <div className="Blog">
                 <header>
@@ -55,15 +16,8 @@ class Blog extends Component {
                         </ul>
                     </nav>
                 </header>
-                <section className="Posts">
-                    {posts}
-                </section>
-                <section>
-                    <FullPost id={this.state.selectedPostId} />
-                </section>
-                <section>
-                    <NewPost />
-                </section>
+                <Route path='/' exact render={()=><h1>Home</h1>} />
+                <Route path='/' render={()=><h1>Home 2</h1>} />
             </div>
         );
     }
